requirements.txt
pandas
numpy
faker
uvicorn
fastapi
requests
snowflake-sqlalchemy
python-dotenv
colorama
llama-index-core
llama-index-llms-openai
mlxtend